Tweak wording.
The wording in the CHANGES.md

  To disable these additional checks when the client connects to
  the database,  `setSSLAuthenticationDisabled(true)` can be called
  on the `ConnectOptions` object passed to the `CloudantClient`
  constructor.

sounds like you can invoke the CloudantClient constructor with a
ConnectOptions object and then afterwards call setSSLAuthenticationDisabled
on it. But this will not have the desired effect, because the
ConnectOptions object is never accessed again by the CloudantClient
class, after the constructor terminated. I would therefore suggest
to change the wording:

  To disable these additional checks when the client connects to
  the database,  `setSSLAuthenticationDisabled(true)` can be called
  on the `ConnectOptions` object before you pass it to the `CloudantClient`
  constructor.