Added client and db ExternalResources
Create JUnit ExternalResource types for both the CloudantClient and Database objects.
This allows before and after methods to be applied at the test method or test class level
 by using the @Rule and @ClassRule annotations respectively, but keeps the setup and
teardown code of these resources in a single place, allowing for easy enforcement of test
best practice (such as using a unique Database name).

BugId: 52580