Remove GRADLE-3112 workaround
Remove the filtering workaround because it breaks other uses of
filtering (e.g. running a single test case). The parameterized tests
still run as part of the unfiltered tasks (test, check,
cloudantServiceTest etc).
A fix is pending for GRADLE-3112 anyway so the ability to filter
parameterized tests can be fixed correctly by a gradle upgrade later.