Fixed service tests for http replacement and cookie auth
Instead of calling execute directly on the HttpConnection the tests should use
the client executeRequest method in UnicodeTest and Utils#removeReplicatorTestDoc.

The HttpProxyTest now asserts that it found the proxy auth header and fixed issue
with test regex for matching header.

The SimpleHttpServer now terminates the final blocking accept to shutdown cleanly.

Removed code for dummy username and password from SSL tests as cookie auth changes
made it unnecessary.

Increased the wait time of the ReplicatorTest and they now throw a TimeoutException
instead of failing an assertion if the timeout is reached.

Added new tests for 1 of either username or password being null.

BugId: 48522