Made saveAttachment implementation honour API/doc intent
Clarified javadoc for the Database.saveAttachment with id & rev method.
Fixed the implementation to generate a UUID instead of throwing an assertion
failure if null is passed as a document ID.

New test for a generating a new document with a generated ID from
saveAttachment: attachmentattachmentStandaloneNullIdNullRev.
New test for generating a new document with a supplied ID and an
attachment: attachmentStandaloneGivenId.

Added additional verification for id and rev combinations, asserting:
If id is null, rev is also null (new create generated ID case).
If id is provided, then it is not an empty string.
If id is provided, then docRev is not an empty string (can be null for
new create or a revision for an ID).
Added new tests for these IllegalArgumentException cases.