Added handling for 429 too many requests
Created RequestLimitInterceptor to check for 429 responses.

Added RequestLimitInterceptor in HttpConnection and wrapping
and unwrapping of IOException.

Added getter for number of retries remaining in HttpConnection to
allow interceptors to change behaviour based on if there are more
retries.

Added tests for 429 cases.

Updated CouchDbClient to throw TooManyRequestsException if a
429 is received and the number of retries was exhausted.

Updated CHANGES.md