Fixed use_index option
Changed to use `JsonElement` instead of `JsonArray` to support both
string and array use cases.

Stopped adding to the array if `useIndex` is called more than once,
replace the property instead.

Added the `use_index` option to the request only in cases where
`useIndex` was called instead of adding an empty array.

Added null checks for parameters to FindByIndexOptions methods.

Added 2 new tests to check the type of the use_index property:
-useIndexDesignDocJsonTypeIsString
-useIndexDesignDocAndIndexNameJsonTypeIsArray

Added a new test to check that the use_index property is not
specified if the option was not used:
-useIndexNotSpecified

Added a test to validate that the use_index property is replaced
if the method is called multiple times on the same FindByIndexOptions:
-useIndexReplaced

Updated CHANGES.md