Closed streams and documented the need to close
There were a couple of places where users of CouchDbUtil were not
closing streams. CouchDbUtil.getAsString was already closing the
streams, so it makes sense if the rest of the util functions that use
streams to do the same and take the onus off of the caller to close the
stream.

Added comments to remind users to close InputStreams in various places.

Moved stream closes into finally blocks to make sure they are called.
* Closed stream in a finally for `streamToString` and removed
  `Scanner` parsing in favour of `IOUtils.toString(InputStream)`.
* Routed response calls in HttpConnection through fewer lines
  and then moved the remaining close into a finally block to
  make sure it is called.