Moved isOkUsable check into own class
The complexity of the OkHttpClientHttpUrlConnectionFactory increased to
the point where it was challenging to continue maintaining it without
inadvertently loading okhttp classes. Such loads prevents okhttp being
optional. To prevent that problem the OkHelper class has only the
isOkUsable check.

Fixed the test mocks to match.
Added a new test that uses a special URLClassLoader with a reflective
load of OkHelper to ensure that classes aren't accidentally loaded.