Use an Interceptor to set the User-Agent header
To make it possible to use java-cloudant and sync-android in the same
application. To do this, the "Library Version" Concept had to be removed.
It is replaced with `UserAgentInterceptor` which will load properties
"user.agent.name" and "user.agent.version" from the the specified properties
file using the provided class loader.