Modified read/connect timeout tests
Removed SleepingDispatcher:
Removed `SleepingDispatcher` class in favour of using MockWebServer’s
`SocketPolicy.NO_RESPONSE`.
Refactored the tests that used it as necessary.

Fixed issue with connection timeout test:
The `connectionTimeout()` test was passing, but taking a lot longer
than it should. Refactored to use the loopback address for that test.

Added read timeout to connection timeout test for openjdk:
Unfortunately openjdk doesn't honour the connect timeout so added a
read timeout as well so that the test doesn't take too long on that
platform.